clear _is_custom_registry
