+++ demisto_sdk/commands/common/tests/docker_helper_test.py
@@ -782,3 +782,23 @@ class TestGetPythonVersionDemistoextendedFallback:
 
         with pytest.raises(Exception, match="docker pull failed"):
             get_python_version("demisto/python3:3.10.11.54799-unique-test")
+
+
+@pytest.mark.parametrize(
+    "registry, expected",
+    [
+        ("gcr.io/xsoar-registry", True),  # bare host
+        ("https://eu.gcr.io/v2/proj", True),  # gcr.io subdomain
+        ("europe-west4-docker.pkg.dev/proj/repo", True),  # pkg.dev
+        ("https://registry-1.docker.io/v2", False),  # non-GAR
+        ("https://gcr.io.evil.com/v2/x", False),  # lookalike suffix
+        ("https://evil.com/gcr.io", False),  # lookalike in path
+    ],
+)
+def test_is_gar_registry_matches_host_not_substring(registry, expected):
+    """
+    _is_gar_registry must match on the parsed host, so lookalike hosts such as
+    gcr.io.evil.com or evil.com/gcr.io are not treated as GAR (CodeQL: incomplete
+    URL substring sanitization).
+    """
+    assert dhelper._is_gar_registry(registry) is expected
